feat(transport): Rewrote transport system and upgraded clientId to 64bit
Changes all ClientIds to 64bit integers and reworks the transport system to be more modern.

BREAKING CHANGE: Changes the transport interface and uses 64bit clientIds

+++ MLAPI-Editor/NetworkedObjectEditor.cs
@@ -57,7 +57,7 @@ namespace UnityEditor
 
                     if (showObservers)
                     {
-                        HashSet<uint>.Enumerator observerClientIds = networkedObject.GetObservers();
+                        HashSet<ulong>.Enumerator observerClientIds = networkedObject.GetObservers();
                     
                         EditorGUI.indentLevel += 1;
