+++ MLAPI/Data/NetworkedCollections/NetworkedDictionary.cs
@@ -313,7 +313,7 @@ namespace MLAPI.NetworkedVar.Collections
         }
 
         /// <inheritdoc />
-        public bool CanClientWrite(uint clientId)
+        public bool CanClientWrite(ulong clientId)
         {
             switch (Settings.WritePermission)
             {
@@ -334,7 +334,7 @@ namespace MLAPI.NetworkedVar.Collections
         }
 
         /// <inheritdoc />
-        public bool CanClientRead(uint clientId)
+        public bool CanClientRead(ulong clientId)
         {
             switch (Settings.ReadPermission)
             {
